bridge/requirements.txt
bleak
paho-mqtt
